changed all integer type naming, added turn-off button

+++ flight_ctrl_v3/calibration.h
@@ -1,11 +1,13 @@
 #ifndef calibration_h
 #define calibration_h
 
+#include <stdint.h>
+
 #include "calibration_struct.h"
 
 void calibration_default(Calibration *);
 void calibration_save(Calibration *);
 void calibration_load(Calibration *);
-void calibration_write(Calibration *, signed long, signed long);
+void calibration_write(Calibration *, int32_t, int32_t);
 
 #endif
