+++ flight_ctrl_v3/timer.h
@@ -1,9 +1,11 @@
 #ifndef timer_h_inc
 #define timer_h_inc
 
+#include <stdint.h>
+
 void timer0_init();
 void timer1_init();
-volatile signed long timer0_elapsed();
-volatile signed long timer1_elapsed();
+volatile int32_t timer0_elapsed();
+volatile int32_t timer1_elapsed();
 
 #endif
